image_scraper/pyproject.toml
[tool.poetry]
name = "image-scraper"
version = "0.1.0"
description = ""
authors = ["sadelcarpio <sadelcarpioa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
scrapy = ">=2.11.0,<2.12.0"
pillow = ">=10.0.1,<10.1.0"
selenium = ">=4.14.0,<4.15.0"
google-cloud-storage = ">=2.12.0,<2.13.0"
scrapyd-client = ">=1.2.3,<1.3.0"
scrapyd = ">=1.4.3,<1.5.0"
confluent-kafka = ">=2.3.0,<2.4.0"
google-cloud-pubsub = "^2.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
